ptrdiff_t instead of int32_t/int64_t pointer value

+++ include/sys/atomic.h
@@ -5,10 +5,14 @@
 #include <Windows.h>
 typedef int		int32_t;
 typedef __int64 int64_t;
+#elif defined(OS_LINUX)
+#include <inttypes.h>
+#include <stdint.h>
 #elif defined(OS_MAC)
 #include <libkern/OSAtomic.h>
 #endif
 
+#include <stddef.h>
 #include <assert.h>
 
 //-------------------------------------------------------------------------------------
@@ -28,28 +32,28 @@ typedef __int64 int64_t;
 #if defined(OS_WINDOWS)
 inline int32_t atomic_increment32(volatile int32_t *value)
 {
-	assert((int32_t)value % 4 == 0);
+	assert((ptrdiff_t)value % 4 == 0);
 	assert(sizeof(LONG) == sizeof(int32_t));
 	return InterlockedIncrement((LONG volatile*)value);
 }
 
 inline int32_t atomic_decrement32(volatile int32_t *value)
 {
-	assert((int32_t)value % 4 == 0);
+	assert((ptrdiff_t)value % 4 == 0);
 	assert(sizeof(LONG) == sizeof(int32_t));
 	return InterlockedDecrement((LONG volatile*)value);
 }
 
 inline int32_t atomic_add32(volatile int32_t *value, int32_t incr)
 {
-	assert((int32_t)value % 4 == 0);
+	assert((ptrdiff_t)value % 4 == 0);
 	assert(sizeof(LONG) == sizeof(int32_t));
 	return InterlockedExchangeAdd((LONG volatile*)value, incr) + incr; // The function returns the initial value of the Addend parameter.
 }
 
 inline int atomic_cas32(volatile int32_t *value, int32_t oldvalue, int32_t newvalue)
 {
-	assert((int32_t)value % 4 == 0);
+	assert((ptrdiff_t)value % 4 == 0);
 	assert(sizeof(LONG) == sizeof(int32_t));
 	return oldvalue == InterlockedCompareExchange((LONG volatile*)value, newvalue, oldvalue) ? 1 : 0;
 }
@@ -57,9 +61,9 @@ inline int atomic_cas32(volatile int32_t *value, int32_t oldvalue, int32_t newva
 inline int atomic_cas_ptr(void* volatile *value, void *oldvalue, void *newvalue)
 {
 #if defined(_WIN64) || defined(WIN64)
-	assert(0 == (int64_t)value % 8 && 0 == (int64_t)oldvalue % 8 && 0 == (int64_t)newvalue % 8);
+	assert(0 == (ptrdiff_t)value % 8 && 0 == (ptrdiff_t)oldvalue % 8 && 0 == (ptrdiff_t)newvalue % 8);
 #else
-	assert(0 == (int32_t)value % 4 && 0 == (int32_t)oldvalue % 4 && 0 == (int32_t)newvalue % 4);
+	assert(0 == (ptrdiff_t)value % 4 && 0 == (ptrdiff_t)oldvalue % 4 && 0 == (ptrdiff_t)newvalue % 4);
 #endif
 	return oldvalue == InterlockedCompareExchangePointer(value, newvalue, oldvalue) ? 1 : 0;
 }
@@ -68,28 +72,28 @@ inline int atomic_cas_ptr(void* volatile *value, void *oldvalue, void *newvalue)
 
 inline int64_t atomic_increment64(volatile int64_t *value)
 {
-	assert((int32_t)value % 8 == 0);
+	assert((ptrdiff_t)value % 8 == 0);
 	assert(sizeof(LONGLONG) == sizeof(int64_t));
 	return InterlockedIncrement64(value);
 }
 
 inline int64_t atomic_decrement64(volatile int64_t *value)
 {
-	assert((int32_t)value % 8 == 0);
+	assert((ptrdiff_t)value % 8 == 0);
 	assert(sizeof(LONGLONG) == sizeof(int64_t));
 	return InterlockedDecrement64(value);
 }
 
 inline int64_t atomic_add64(volatile int64_t *value, int64_t incr)
 {
-	assert((int32_t)value % 8 == 0);
+	assert((ptrdiff_t)value % 8 == 0);
 	assert(sizeof(LONGLONG) == sizeof(int64_t));
 	return InterlockedExchangeAdd64(value, incr) + incr; // The function returns the initial value of the Addend parameter.
 }
 
 inline int atomic_cas64(volatile int64_t *value, int64_t oldvalue, int64_t newvalue)
 {
-	assert((int32_t)value % 8 == 0);
+	assert((ptrdiff_t)value % 8 == 0);
 	assert(sizeof(LONGLONG) == sizeof(int64_t));
 	return oldvalue == InterlockedCompareExchange64(value, newvalue, oldvalue) ? 1 : 0;
 }
@@ -100,58 +104,58 @@ inline int atomic_cas64(volatile int64_t *value, int64_t oldvalue, int64_t newva
 #elif defined(OS_MAC)
 inline int32_t atomic_increment32(volatile int32_t *value)
 {
-	assert((int32_t)value % 4 == 0);
+	assert((ptrdiff_t)value % 4 == 0);
 	return OSAtomicIncrement32(value);
 }
 
 inline int64_t atomic_increment64(volatile int64_t *value)
 {
-	assert((int32_t)value % 8 == 0);
+	assert((ptrdiff_t)value % 8 == 0);
 	return OSAtomicIncrement64(value);
 }
 
 inline int32_t atomic_decrement32(volatile int32_t *value)
 {
-	assert((int32_t)value % 4 == 0);
+	assert((ptrdiff_t)value % 4 == 0);
 	return OSAtomicDecrement32(value);
 }
 
 inline int64_t atomic_decrement64(volatile int64_t *value)
 {
-	assert((int32_t)value % 8 == 0);
+	assert((ptrdiff_t)value % 8 == 0);
 	return OSAtomicDecrement64(value);
 }
 
 inline int32_t atomic_add32(volatile int32_t *value, int32_t incr)
 {
-	assert((int32_t)value % 4 == 0);
+	assert((ptrdiff_t)value % 4 == 0);
 	return OSAtomicAdd32(incr, value);
 }
 
 inline int64_t atomic_add64(volatile int64_t *value, int64_t incr)
 {
-	assert((int32_t)value % 8 == 0);
+	assert((ptrdiff_t)value % 8 == 0);
 	return OSAtomicAdd64(incr, value);
 }
 
 inline int atomic_cas32(volatile int32_t *value, int32_t oldvalue, int32_t newvalue)
 {
-	assert((int32_t)value % 4 == 0);
+	assert((ptrdiff_t)value % 4 == 0);
 	return OSAtomicCompareAndSwap32(oldvalue, newvalue, value) ? 1 : 0;
 }
 
 inline int atomic_cas64(volatile int64_t *value, int64_t oldvalue, int64_t newvalue)
 {
-	assert((int32_t)value % 8 == 0);
+	assert((ptrdiff_t)value % 8 == 0);
 	return OSAtomicCompareAndSwap64(oldvalue, newvalue, value) ? 1 : 0;
 }
 
 inline int atomic_cas_ptr(void* volatile *value, void *oldvalue, void *newvalue)
 {
 #if TARGET_CPU_X86_64 || TARGET_CPU_PPC64
-	assert(0 == (int32_t)value % 8 && 0 == (int32_t)oldvalue % 8 && 0 == (int32_t)newvalue % 8);
+	assert(0 == (ptrdiff_t)value % 8 && 0 == (ptrdiff_t)oldvalue % 8 && 0 == (ptrdiff_t)newvalue % 8);
 #else
-	assert(0 == (int32_t)value % 4 && 0 == (int32_t)oldvalue % 4 && 0 == (int32_t)newvalue % 4);
+	assert(0 == (ptrdiff_t)value % 4 && 0 == (ptrdiff_t)oldvalue % 4 && 0 == (ptrdiff_t)newvalue % 4);
 #endif
 	return OSAtomicCompareAndSwapPtr(oldvalue, newvalue, value) ? 1 : 0;
 }
@@ -166,55 +170,55 @@ inline int atomic_cas_ptr(void* volatile *value, void *oldvalue, void *newvalue)
 #if __GNUC__>=4 && __GNUC_MINOR__>=1
 inline int32_t atomic_increment32(volatile int32_t *value)
 {
-	assert((long)value % 4 == 0);
+	assert((ptrdiff_t)value % 4 == 0);
 	return __sync_add_and_fetch_4(value, 1);
 }
 
 inline int32_t atomic_decrement32(volatile int32_t *value)
 {
-	assert((long)value % 4 == 0);
+	assert((ptrdiff_t)value % 4 == 0);
 	return __sync_sub_and_fetch_4(value, 1);
 }
 
 inline int32_t atomic_add32(volatile int32_t *value, int32_t incr)
 {
-	assert((long)value % 4 == 0);
+	assert((ptrdiff_t)value % 4 == 0);
 	return __sync_add_and_fetch_4(value, incr);
 }
 
 inline int atomic_cas32(volatile int32_t *value, int32_t oldvalue, int32_t newvalue)
 {
-	assert((long)value % 4 == 0);
+	assert((ptrdiff_t)value % 4 == 0);
 	return __sync_bool_compare_and_swap_4(value, oldvalue, newvalue) ? 1 : 0;
 }
 
 inline int atomic_cas_ptr(void* volatile *value, void *oldvalue, void *newvalue)
 {
-	assert(0 == (long)value % 4 && 0 == (long)oldvalue % 4 && 0 == (long)newvalue % 4);
+	assert(0 == (ptrdiff_t)value % 4 && 0 == (ptrdiff_t)oldvalue % 4 && 0 == (ptrdiff_t)newvalue % 4);
 	return __sync_bool_compare_and_swap(value, oldvalue, newvalue);
 }
 
 inline int64_t atomic_increment64(volatile int64_t *value)
 {
-	assert((long)value % 8 == 0);
+	assert((ptrdiff_t)value % 8 == 0);
 	return __sync_add_and_fetch_8(value, 1);
 }
 
 inline int64_t atomic_decrement64(volatile int64_t *value)
 {
-	assert((long)value % 8 == 0);
+	assert((ptrdiff_t)value % 8 == 0);
 	return __sync_sub_and_fetch_8(value, 1);
 }
 
 inline int64_t atomic_add64(volatile int64_t *value, int64_t incr)
 {
-	assert((long)value % 8 == 0);
+	assert((ptrdiff_t)value % 8 == 0);
 	return __sync_add_and_fetch_8(value, incr);
 }
 
 inline int atomic_cas64(volatile int64_t *value, int64_t oldvalue, int64_t newvalue)
 {
-	assert((long)value % 8 == 0);
+	assert((ptrdiff_t)value % 8 == 0);
 	return __sync_bool_compare_and_swap_8(value, oldvalue, newvalue) ? 1 : 0;
 }
 
@@ -223,7 +227,7 @@ inline int atomic_cas64(volatile int64_t *value, int64_t oldvalue, int64_t newva
 
 inline int32_t atomic_add32(volatile int32_t *value, int32_t incr)
 {
-	assert((int32_t)value % 4 == 0);
+	assert((ptrdiff_t)value % 4 == 0);
 	int a, b, c;
 	asm volatile(	"0:\n\t"
 					"ldr %0, [%3]\n\t"
@@ -240,13 +244,13 @@ inline int32_t atomic_add32(volatile int32_t *value, int32_t incr)
 
 inline int32_t atomic_increment32(volatile int32_t *value)
 {
-	assert((int32_t)value % 4 == 0);
+	assert((ptrdiff_t)value % 4 == 0);
 	return atomic_add32(value, 1);
 }
 
 inline int32_t atomic_decrement32(volatile int32_t *value)
 {
-	assert((int32_t)value % 4 == 0);
+	assert((ptrdiff_t)value % 4 == 0);
 	return atomic_add32(value, -1);
 }
 
@@ -255,7 +259,7 @@ inline int32_t atomic_decrement32(volatile int32_t *value)
 inline int32_t atomic_add32(volatile int32_t *value, int32_t incr)
 {
 	int32_t r = incr;
-	assert((long)value % 4 == 0);
+	assert((ptrdiff_t)value % 4 == 0);
 	asm volatile ("lock; xaddl %0,%1"
 		: "+r" (r), "+m" (*value));
 	return r + incr;
@@ -263,20 +267,20 @@ inline int32_t atomic_add32(volatile int32_t *value, int32_t incr)
 
 inline int32_t atomic_increment32(volatile int32_t *value)
 {
-	assert((long)value % 4 == 0);
+	assert((ptrdiff_t)value % 4 == 0);
 	return atomic_add32(value, 1);
 }
 
 inline int32_t atomic_decrement32(volatile int32_t *value)
 {
-	assert((long)value % 4 == 0);
+	assert((ptrdiff_t)value % 4 == 0);
 	return atomic_add32(value, -1);
 }
 
 inline int atomic_cas32(volatile int32_t *value, int32_t oldvalue, int32_t newvalue)
 {
 	int32_t prev = oldvalue;
-	assert((long)value % 4 == 0);
+	assert((ptrdiff_t)value % 4 == 0);
 	asm volatile ("lock; cmpxchgl %2, %1"
 		: "+a" (prev), "+m" (*value)
 		: "r" (newvalue));
@@ -287,7 +291,7 @@ inline int atomic_cas64(volatile int64_t *value, int64_t oldvalue, int64_t newva
 {
 	int64_t prev = oldvalue;
 #if defined(__LP64__)
-	assert((long)value % 8 == 0);
+	assert((ptrdiff_t)value % 8 == 0);
 	asm volatile ("lock ; cmpxchgq %2, %1"
 		: "+a" (prev), "+m" (*value)
 		: "q" (newvalue));
@@ -312,7 +316,7 @@ inline int64_t atomic_add64(volatile int64_t *value, int64_t incr)
 {
 #if defined(__LP64__)
 	int64_t r = incr;
-	assert((long)value % 8 == 0);
+	assert((ptrdiff_t)value % 8 == 0);
 	asm volatile ("lock; xaddq %0,%1"
 		: "+r" (r), "+m" (*value));
 	return r + incr;
@@ -328,13 +332,13 @@ inline int64_t atomic_add64(volatile int64_t *value, int64_t incr)
 
 inline int64_t atomic_increment64(volatile int64_t *value)
 {
-	assert((long)value % 8 == 0);
+	assert((ptrdiff_t)value % 8 == 0);
 	return atomic_add64(value, 1);
 }
 
 inline int64_t atomic_decrement64(volatile int64_t *value)
 {
-	assert((long)value % 8 == 0);
+	assert((ptrdiff_t)value % 8 == 0);
 	return atomic_add64(value, -1);
 }
